new "invoked" event
We trigger invoked" event on every invocation the server makes for client methods. This provides the user a way to be notified of any call request by server, whether or not he has defined a client method for what the server has asked. IMHO this is useful for debugging. If by any reason, the client code misses a method the server calls, the developer will find the error quickly.